bugfix in toc.xqm:
toc mets:fptrs representing resourcefragments got lost, when preceding or following a mets:div because the generated stylesheet decided outputting either nested divs or fptrs: a structure like

<TEI>
   <pb/>
   ...
   <div type="toBePutIntoToc">...</div>
   ...
   <pb/>
</TEI>

would have only resulted in

<mets:div TPY="resource">
    <mets:div TYPE="toBePutIntoToc">
        <mets:fptr/>
    </mets:div>
</mets:div>

... thus loosing the representation of the preceding and following resourcefragments/pages.
The stylesheet's behaviour is changed now to:

<mets:div TYPE="resource">
    <mets:fptr>
        <mets:area/>
    <mets:fptr>
    <mets:div TYPE="toBePutIntoToc">
        <mets:fptr/>
    </mets:div>
    <mets:fptr>
        <mets:area/>
    <mets:fptr>
</mets:div>

HOWEVER the resulting structure is not valid against the METS schema: while there might be a fptr BEFORE a div,
fptrs AFTER a div are not allowed. TODO: find a solution, maybe auto-wrap those floating fptrs into divs that are
ignored in the output of fcs.toc - losing fptr information is not an option.